# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Musiyo API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/elements": {
      "get": {
        "summary": "List Elements",
        "operationId": "list_elements_api_v1_elements_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Element"
                  },
                  "type": "array",
                  "title": "Response List Elements Api V1 Elements Get"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories": {
      "get": {
        "summary": "List Categories",
        "operationId": "list_categories_api_v1_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogFacet"
                  },
                  "type": "array",
                  "title": "Response List Categories Api V1 Categories Get"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections": {
      "get": {
        "summary": "List Collections",
        "operationId": "list_collections_api_v1_collections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogFacet"
                  },
                  "type": "array",
                  "title": "Response List Collections Api V1 Collections Get"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog": {
      "get": {
        "summary": "Get Catalog",
        "operationId": "get_catalog_api_v1_catalog_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Query"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "collection",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Collection"
            }
          },
          {
            "name": "highlighted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Highlighted"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 24,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "schema_version",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Schema Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogPage"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/elements/{element_id}": {
      "get": {
        "summary": "Get Element",
        "operationId": "get_element_api_v1_elements__element_id__get",
        "parameters": [
          {
            "name": "element_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Element Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Element"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/elements/{element_id}/resources/{resource_id}": {
      "get": {
        "summary": "Get Resource",
        "operationId": "get_resource_api_v1_elements__element_id__resources__resource_id__get",
        "parameters": [
          {
            "name": "element_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Element Id"
            }
          },
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/tours/{tour_id}": {
      "get": {
        "summary": "Get Tour",
        "operationId": "get_tour_api_v1_tours__tour_id__get",
        "parameters": [
          {
            "name": "tour_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tour Id"
            }
          },
          {
            "name": "schema_version",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Schema Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tour"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {},
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CatalogFacet": {
        "properties": {
          "slug": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "element_count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Element Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "slug",
          "name",
          "element_count"
        ],
        "title": "CatalogFacet"
      },
      "CatalogItem": {
        "properties": {
          "slug": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "community": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Community"
          },
          "thumbnail_resource_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "pattern": "^\\S+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Resource Id"
          },
          "has_3d_model": {
            "type": "boolean",
            "title": "Has 3D Model",
            "default": false
          },
          "has_narration": {
            "type": "boolean",
            "title": "Has Narration",
            "default": false
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NamedTerm"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "slug",
          "title",
          "description"
        ],
        "title": "CatalogItem"
      },
      "CatalogPage": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "CatalogPage"
      },
      "Community": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "people": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "People"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "people"
        ],
        "title": "Community"
      },
      "Credit": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Credit"
      },
      "Element": {
        "properties": {
          "slug": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Slug"
          },
          "language": {
            "type": "string",
            "const": "es",
            "title": "Language",
            "default": "es"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "community": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Community"
              },
              {
                "type": "null"
              }
            ]
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NamedTerm"
              },
              {
                "type": "null"
              }
            ]
          },
          "collections": {
            "items": {
              "$ref": "#/components/schemas/NamedTerm"
            },
            "type": "array",
            "title": "Collections"
          },
          "technique": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technique"
          },
          "materials": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Materials"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/TextBlock"
            },
            "type": "array",
            "title": "Blocks"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          },
          "credits": {
            "items": {
              "$ref": "#/components/schemas/Credit"
            },
            "type": "array",
            "title": "Credits"
          },
          "restrictions": {
            "items": {
              "$ref": "#/components/schemas/Restriction"
            },
            "type": "array",
            "title": "Restrictions"
          },
          "resources": {
            "items": {
              "$ref": "#/components/schemas/Resource"
            },
            "type": "array",
            "title": "Resources"
          },
          "last_modified": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LastModified"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "slug",
          "title",
          "description"
        ],
        "title": "Element"
      },
...
      "LastModified": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "author": {
            "type": "string",
            "title": "Author"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "author"
        ],
        "title": "LastModified"
      },
      "NamedTerm": {
        "properties": {
          "slug": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "slug",
          "name"
        ],
        "title": "NamedTerm"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "code": {
            "type": "string",
            "enum": [
              "not_found",
              "schema_incompatible",
              "validation"
            ],
            "title": "Code"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail"
        ],
        "title": "Problem"
      },
      "Resource": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "model_3d",
              "image",
              "audio",
              "video",
              "narration",
              "subtitles"
            ],
            "title": "Kind"
          },
          "mime": {
            "type": "string",
            "title": "Mime"
          },
          "byte_count": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Byte Count"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "alternative_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alternative Text"
          },
          "credit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit"
          },
          "provenance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/ResourceVariant"
            },
            "type": "array",
            "title": "Variants"
          },
          "transcription": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcription"
          },
          "subtitles_resource_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "pattern": "^\\S+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitles Resource Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "mime"
        ],
        "title": "Resource"
      },
...
      "Restriction": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "no_download",
              "no_reuse",
              "other"
            ],
            "title": "Kind"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "description"
        ],
        "title": "Restriction"
      },
...
      "Source": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "publication",
              "interview",
              "other"
            ],
            "title": "Kind"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "reference"
        ],
        "title": "Source"
      },
      "TextBlock": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "documented_fact",
              "testimony",
              "interpretation"
            ],
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "pattern": "^\\S+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribution"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "text"
        ],
        "title": "TextBlock"
      },
      "Tour": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version"
          },
          "tour": {
            "$ref": "#/components/schemas/TourMetadata"
          },
          "rooms": {
            "items": {
              "$ref": "#/components/schemas/Room"
            },
            "type": "array",
            "minItems": 1,
            "title": "Rooms"
          },
          "guide": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Guide"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "schema_version",
          "tour",
          "rooms"
        ],
        "title": "Tour"
      },
      "TourMetadata": {
        "properties": {
          "key": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "revision": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "name"
        ],
        "title": "TourMetadata"
      }
    }
  }
}